backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Firebase (optional - for production)
# firebase-admin==6.2.0

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0

